refactor: AccountConnection now handle OAuth all by itself :art:
