+++ src/components/services/CreateAccountService.jsx
@@ -3,14 +3,13 @@ import React from 'react'
 import AccountConnection from '../../containers/AccountConnection'
 
 const CreateAccountService = (props) => {
-  const { konnector, onSubmit, onOAuth } = props
+  const { konnector, onSubmit } = props
   return (
     <div className='coz-service-content'>
       <AccountConnection
         connector={konnector}
         fields={konnector.fields}
         onSubmit={onSubmit}
-        onOAuth={onOAuth}
         {...props}
         />
     </div>
